flink-job/requirements.txt
apache-flink==1.18.0
kafka-python==2.0.2
pymongo==4.6.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
requests==2.31.0
regex==2023.10.3